MovieList and MovieDetails networking feature is created

+++ app/src/main/java/com/mycodeflow/data/Genre.kt
@@ -1,3 +1,6 @@
 package com.mycodeflow.data
 
-data class Genre(val id: Int, val name: String)
+data class Genre (
+    val id: Long,
+    val name: String
+)
